# backend/docs/swagger.yaml
basePath: /v1/api/radar-hub-manager
definitions:
...
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName:
    enum:
    - ADMIN
    - OPERATOR
    - HQ
    type: string
    x-enum-comments:
      RoleAdmin: Quản trị hệ thống
      RoleHQ: Cán bộ Sở chỉ huy
      RoleOperator: Nhân viên vận hành trạm
    x-enum-descriptions:
    - Quản trị hệ thống
    - Nhân viên vận hành trạm
    - Cán bộ Sở chỉ huy
    x-enum-varnames:
    - RoleAdmin
    - RoleOperator
    - RoleHQ
...
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station:
    properties:
      created_at:
        type: integer
      distance_to_coast:
        description: km
        type: number
      elevation:
        description: mét so với mực nước biển
        type: number
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      note:
        description: ghi chú tự do
        type: string
      status:
        description: '"ACTIVE" / "INACTIVE" (tính từ lịch)'
        type: string
      updated_at:
        type: integer
    type: object
  github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User:
    properties:
      created_at:
        type: integer
      full_name:
        type: string
      id:
        description: UUID dưới dạng chuỗi
        type: integer
      last_login:
        type: integer
      password:
        description: Chỉ dùng khi tạo mới, không trả về
        type: string
      role_id:
        $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName'
      station:
        $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.Station'
      station_id:
        description: 'Nếu là OPERATOR: trạm mà người dùng trực thuộc'
        type: integer
      updated_at:
        type: integer
      username:
        type: string
    type: object
...
  internal_handlers.CreateUserRequest:
    properties:
      full_name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
      role_id:
        allOf:
        - $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName'
        example: OPERATOR
      station_id:
        example: 1
        type: integer
      username:
        example: johndoe
        type: string
    required:
    - full_name
    - password
    - role_id
    - username
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.UpdateUserRequest:
    properties:
      full_name:
        example: John Smith
        type: string
      password:
        example: newpassword123
        type: string
      role_id:
        allOf:
        - $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.RoleName'
        example: ADMIN
      station_id:
        example: 2
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Radar Hub Manager system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Radar Hub Manager API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all users. Only users with ADMIN role can perform
        this action.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all users (Admin only)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Conflict - Username already exists
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new user (Admin only)
      tags:
      - users
  /users/{username}:
    delete:
      consumes:
      - application/json
      description: Delete a user account. Only users with ADMIN role can perform this
        action.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a user (Admin only)
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user information by username. Only users with ADMIN role can
        perform this action.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user by username (Admin only)
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information. Only users with ADMIN role can perform
        this action.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_lehaisonagentai2_radar-hub-manager_backend_internal_models.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing user (Admin only)
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"